Finished INSERT SQL Method
