+++ core/base/model/BaseModel.php
@@ -7,7 +7,7 @@ namespace core\base\model;
 use core\base\controller\Singleton;
 use core\base\exceptions\DbException;
 
-class BaseModel
+class BaseModel extends BaseModelMethods
 {
 
     use Singleton;
@@ -26,6 +26,14 @@ class BaseModel
         $this->db->query("SET NAMES UTF8"  );
 
     }
+
+    /**
+     * @param $query
+     * @param string $crud = r - SELECT / u - UPDATE / c - INSERT / D - delete
+     * @param string $return_id
+     * @return array|bool|mixed
+     * @throws DbException
+     */
 
     final public function query($query, $crud = 'r', $return_id = 'false' )
     {
@@ -127,214 +135,36 @@ class BaseModel
         return $this->query($query);
     }
 
-    protected function createFields($set, $table = false) {
 
-        $set['fields'] = (is_array($set['fields']) && !empty($set['fields']))
-                            ? $set['fields'] : ['*'];
+    /**
+     * @param $table - таблица для добавления данных
+     * @param array $set - массив параметров
+     * fields => (поле => значение); если не указан, обрабатывается $_POST(поле => значение)
+     * разрешена передача функций (например NOW()) в качестве MySQL функции, обычно строкой
+     * files => [поле => значение]; можно подать массив вида (поле => [массив значений])
+     * except => ['исключение 1', 'исключение 2'] - исключает элементы массива из добавления в запрос
+     * return_id => true | false - возвращать или нет идентификатор вставленной записи
+     * @return  mixed
+     */
 
-        $table = $table ? $table . '.' : '';
+    final public function add($table, $set) {
 
-        $fields = '';
+        $set['fields'] = (is_array($set['fields']) & !empty($set['fields'])) ? $set['fields'] : false;
+        $set['files'] = (is_array($set['files']) & !empty($set['files'])) ? $set['files'] : false;
+        $set['return_id'] = $set['return_id'] ? true : false;
+        $set['except'] = (is_array($set['except']) && !empty($set['except'])) ? $set['except'] : false;
 
+        $insert_arr = $this->createInsert($set['fields'], $set['files'], $set['except']);
 
-        foreach($set['fields'] as $field) {
-            $fields .= $table . $field . ',';
+        if($insert_arr) {
+            $query = "INSERT INTO $table ({$insert_arr['fields']}) VALUES ({$insert_arr['values']})";
+            return $this->query($query, 'c', $set['return_id']);
         }
 
-        return $fields;
-    }
-
-    protected function createOrder($set, $table = false) {
-
-        $table = $table ? $table . '.' : '';
-
-        $order_by = '';
-
-        if(is_array($set['order']) && !empty($set['order'])) {
-
-            $set['order_direction'] = (is_array($set['order_direction']) && !empty($set['order_direction']))
-                ? $set['order_direction'] : ['ASC'];
-
-            $order_by = 'ORDER BY ';
-            $direct_count = 0;
-
-            foreach ($set['order'] as $order) {
-                if($set['order_direction'][$direct_count]) {
-                    $order_direction = strtoupper($set['order_direction'][$direct_count]);
-                    $direct_count++;
-                } else {
-                    $order_direction = strtoupper($set['order_direction'][$direct_count - 1]);
-                }
-
-                if(is_int($order)) $order_by .= $order . ' ' . $order_direction . ',';
-                    else $order_by .= $table . $order . ' ' . $order_direction . ',';
-            }
-
-            $order_by = rtrim($order_by, ',');
-
-        }
-
-        return $order_by;
+        return false;
 
     }
 
-    protected function createWhere($set, $table = false, $instruction = 'WHERE ') {
 
-        $table = $table ? $table . '.' : '';
-
-        $where  = '';
-
-        if(is_array($set['where']) && !empty($set['where'])) {
-            $set['operand'] = (is_array($set['operand']) && !empty($set['operand']))
-                ? $set['operand']
-                : ['='];
-            $set['condition'] = (is_array($set['condition']) && !empty($set['condition']))
-                ? $set['condition']
-                : ['AND'];
-            $where  = $instruction;
-
-            $o_count = 0;
-            $c_count = 0;
-
-            foreach ($set['where'] as $key => $item) {
-                $where .= ' ';
-
-                if($set['operand'][$o_count]) {
-                    $operand = $set['operand'][$o_count];
-                    $o_count++;
-                } else {
-                    $operand = $set['operand'][$o_count - 1];
-                }
-
-                if($set['condition'][$c_count]) {
-                    $condition = $set['condition'][$c_count];
-                    $c_count++;
-                } else {
-                    $condition = $set['condition'][$c_count - 1];
-                }
-
-                if($operand === 'IN' || $operand === 'NOT IN') {
-
-                    if(is_string($item) && strpos($item,  'SELECT') === 0) {
-                        $in_str = $item;
-                    } else {
-                        if(is_array($item)) $temp_item = $item;
-                        else $temp_item = explode(',', $item);
-
-                        $in_str = '';
-
-                        foreach ($temp_item as $v) {
-                            $in_str .= "'" . addslashes(trim($v)) . "',";
-                        }
-                    }
-
-                    $where .= $table . $key . ' '. $operand . ' (' . rtrim($in_str,',') . ') ' . $condition;
-
-
-                } elseif (strpos($operand, 'LIKE') !== false ) {
-
-                    $like_template = explode('%', $operand);
-
-                    foreach($like_template as $lt_key => $lt) {
-                        if(!$lt){
-                            if(!$lt_key) {
-                                $item = '%' . $item;
-                            } else {
-                                $item .= '%';
-                            }
-                        }
-                    }
-
-                    $where .= $table . $key . ' LIKE ' . "'" . addslashes($item) . "' " . $condition;
-
-                } else {
-                    if(strpos($item, 'SELECT') === 0) {
-                        $where .= $table . $key . $operand . '(' . $item . ") $condition";
-                    } else {
-                        $where .= $table . $key . $operand . "'" . addslashes($item) . "' $condition";
-                    }
-                }
-
-            }
-
-            $where = substr($where, 0, strrpos($where, $condition));
-        }
-
-        return $where;
-
-    }
-
-    protected function createJoin($set, $table, $new_where = false)
-    {
-        $fields = '';
-        $join = '';
-        $where = '';
-
-        if ($set['join']) {
-
-            $join_table = $table;
-
-            foreach ($set['join'] as $key => $item) {
-
-                if (is_int($key)) {
-                    if (!$item['table']) continue;
-                    else $key = $item['table'];
-                }
-
-                if ($join) $join .= ' ';
-
-                if ($item['on']) {
-
-                    $join_fields = [];
-
-                    switch (2) {
-
-                        case count($item['on']['fields']):
-                            $join_fields = $item['on']['fields'];
-                            break;
-
-                        case count($item['on']):
-                            $join_fields = $item['on'];
-                            break;
-
-                        default:
-
-                            continue 2;
-                            break;
-                    }
-
-                    if (!$item['type']) $join .= 'LEFT JOIN ';
-                    else $join .= trim(strtoupper($item['type'])) . ' JOIN ';
-
-                    $join .= $key . ' ON ';
-
-                    if ($item['on']['table']) $join .= $item['on']['table'];
-                        else $join .= $join_table;
-
-                    $join .= '.' . $join_fields[0] . '=' . $key . '.' . $join_fields[1];
-
-                    $join_table = $key;
-
-                    if ($new_where) {
-
-                        if ($item['where']) {
-                            $new_where = false;
-                        }
-                        $group_condition = 'WHERE';
-                    } else {
-                        $group_condition = $item['group_condition']
-                            ? strtoupper($item['group_condition'])
-                            : 'AND';
-                    }
-
-                    $fields .= $this->createFields($item, $key);
-                    $where .= $this->createWhere($item, $key, $group_condition);
-                }
-            }
-        }
-
-        return compact('fields', 'join', 'where');
-
-    }
 
 }
